Add help files to replace the help text in cli.cpp
I did this for smaller size and so that I can edit a text file instead of having
to edit the output in the form of c++ strings.

CMakeLists.txt now sets the mode to debug if not specified.
It will also create a symlink for the help files in debug mode, so I can edit
without having to recompile.
It will copy instead of symlink in any non-debug mode