Improve error handling
Add 2 new structs, Result<T> and SimpleResult that have are used to return a
value, error or warning, and an enum that says what it is.

Add more error messages throughout to catch any c++ errors that we don't want
the user to see. There still might be a few left, though.

Use ChatGPT to generate a ctest file because I'm too lazy to do it myself.